# Allow SparseMatrixCSC in Constant