--- a/src/solution.jl
+++ b/src/solution.jl
@@ -85,9 +85,11 @@ function solve!(
     if problem_vexity(p) in (ConcaveVexity(), NotDcp())
         throw(DCPViolationError())
     end
-    context, formulation_time, formulation_bytes =
-        @timed Context(p, optimizer_factory)
-    # @show (formulation_time, Base.format_bytes(formulation_bytes))
+    context, (stats...) = @timed Context(p, optimizer_factory)
+    if !silent_solver
+        s = round(stats.time; digits = 2), Base.format_bytes(stats.bytes)
+        @info "[Convex.jl] Compilation finished: $(s[1]) seconds, $(s[2]) of memory allocated"
+    end
     if silent_solver
         MOI.set(context.model, MOI.Silent(), true)
     end