--- a/test/test_utilities.jl
+++ b/test/test_utilities.jl
@@ -1295,6 +1295,19 @@ function test_broadcasting()
     return
 end
 
+function test_matrix_constants()
+    I, V = [1, 3, 4], [2.1, 2.2, 3.3]
+    x = SparseArrays.sparsevec(I, V)
+    y = SparseArrays.sparse(I, [1, 1, 1], V, 4, 1)
+    c = constant(x)
+    @test c.value isa SparseArrays.SparseMatrixCSC
+    @test c.value == y
+    c = constant(y)
+    @test c.value isa SparseArrays.SparseMatrixCSC
+    @test c.value == y
+    return
+end
+
 end  # TestUtilities
 
 TestUtilities.runtests()