@ericphanson 2024-05-06
btw, it was my first time profiling end-to-end with a pure julia solver, it's pretty neat to do a `@profview` in vscode and jump to `kkt_update!` and stuff like that!